Magpie models refactoring
