+++ nemo/collections/tts/models/__init__.py
@@ -16,6 +16,7 @@ from nemo.collections.tts.models.aligner import AlignerModel
 from nemo.collections.tts.models.audio_codec import AudioCodecModel
 from nemo.collections.tts.models.base_magpietts import BaseMagpieTTSModel
 from nemo.collections.tts.models.easy_magpietts import EasyMagpieTTSModel
+from nemo.collections.tts.models.easy_magpietts_inference import EasyMagpieTTSInferenceModel
 from nemo.collections.tts.models.easy_magpietts_preference_optimization import EasyMagpieTTSModelOnlinePO
 from nemo.collections.tts.models.fastpitch import FastPitchModel
 from nemo.collections.tts.models.fastpitch_ssl import FastPitchModel_SSL
@@ -39,6 +40,7 @@ __all__ = [
     "InferBatchOutput",
     "MagpieTTSModel",
     "EasyMagpieTTSModel",
+    "EasyMagpieTTSInferenceModel",
     "EasyMagpieTTSModelOnlinePO",
     "MagpieTTSModelOfflinePODataGen",
     "MagpieTTSModelOfflinePO",
